Initial 64bit

+++ patches/tModLoader/Terraria.ModLoader.IO/WorldIO.cs
@@ -616,13 +616,16 @@ namespace Terraria.ModLoader.IO
 			path = Path.ChangeExtension(path, ".twld");
 			if (!cloudSave)
 			{
-#if WINDOWS
-				FileOperationAPIWrapper.MoveToRecycleBin(path);
-				FileOperationAPIWrapper.MoveToRecycleBin(path + ".bak");
-#else
-				File.Delete(path);
-				File.Delete(path + ".bak");
-#endif
+				if(PlatformUtilities.IsWindows)
+				{
+					FileOperationAPIWrapper.MoveToRecycleBin(path);
+					FileOperationAPIWrapper.MoveToRecycleBin(path + ".bak");
+				}
+				else
+				{
+					File.Delete(path);
+					File.Delete(path + ".bak");
+				}
 			}
 			else if (SocialAPI.Cloud != null)
 			{
